There are probably more buggy zero checks in the program, we have to check for zero in the datatype which the value should be, not what it happens to be mid computation. This issue is that for overflow we rely on bytes and shorts being expanded to ints automatically.

+++ generator/opcodes/OpcodeImplementation.cs
@@ -392,7 +392,7 @@ namespace generator
             Registers.Mark(Flag.NN);
             var sum = lhs + rhs;
 
-            Registers.Set(Flag.Z, sum == 0);
+            Registers.Set(Flag.Z, ((byte)sum) == 0);
             Registers.Set(Flag.C, sum > 0xff);
             Registers.Set(Flag.H, lhs.IsHalfCarryAdd(rhs));
             return (byte)sum;
